fix another GCC warning introduced in last commit

+++ src/liblinux/pngrw.c
@@ -76,14 +76,14 @@ void user_read_data(png_structp png_read, png_bytep data, png_size_t length)
 {
     FILE *fPtr = (FILE *) png_get_io_ptr(png_read);
     if (fread(data, 1, length, fPtr) != length)
-        fprintf(stderr, "Failed to read %i bytes from PNG file.\n", length);
+        fprintf(stderr, "Failed to read %i bytes from PNG file.\n", (int) length);
 }
 
 void user_write_data(png_structp png_write, png_bytep data, png_size_t length)
 {
     FILE *fPtr = (FILE *) png_get_io_ptr(png_write);
     if (fwrite(data, 1, length, fPtr) != length)
-        fprintf(stderr, "Failed to write %i bytes to PNG file.\n", length);
+        fprintf(stderr, "Failed to write %i bytes to PNG file.\n", (int) length);
 }
 
 void user_flush_data(png_structp png_read)
